SIMSBIOHUB-61: Rearrange interfaces
