+++ app/src/hooks/api/usePublishApi.ts
@@ -1,6 +1,5 @@
 import { AxiosInstance } from 'axios';
-import { IProjectSubmitForm } from 'components/publish/PublishProjectSections';
-import { ISurveySubmitForm } from 'components/publish/PublishSurveySections';
+import { IProjectSubmitForm, ISurveySubmitForm } from 'interfaces/usePublishApi.interface';
 
 /**
  * Returns a list of all resources
